feat: move solana unwrap tip to swap button
